MessageBag instance was not being returned (Updated)
Calling getMessageBag(), errors(), or messages() was not returning an instance if passes() or fails() was not first called. This way these function will always return a MessageBag instance regardless.

Adhered to L4 syntax standards. Removed the line from getMessageBag() as it wasn't need because it calls messages() anyways.

Thank you so much!