Add the ability to register subscribers
To use this you would simply add the class names you wish to register in the `$subscribe` property in the `App\Providers\EventServiceProvider` class. I've added the property to the class itself because I didn't want to make laravel/laravel dependent on this change. I also kept `$subscribe` singular to remain consistent with `$listen`.

