Improved exception reporting
It can be very useful to have the entire exception passed through to the
logger. An example use case would be logging to bugsnag by having a
bugsnag logger object, and passing the entire exception means we can
extract more detail from it. This is not a breaking change because psr
says that the "message" should be a string, or be stringish, which an
exception object is.