Add methods defined in Gate implementation to contract
There are four methods (before, allows, denies and forUser)
mentioned in the documentation which are defined in the concrete
implementation of Gate but not in the contract.

This leads to issues when type hinting on the Gate interface as
methods are not defined in the type hinted type object:

    public function update(GateContract $gate)
    {
        $gate->allows('post-user');
    }

This either limits the implementation being used to the provided
one or forces the use of the check method when type hinting the
interface which is often less clear than the explicit allows /
denies methods.

There are other methods which are public but not on the contract
but it appears as though they are intended to be internal and as
such should be marked protected. This change does not address
these methods.

This change adds the four documented methods in the public api
to the Gate contract.