Drastically reduce overhead in Blade comment matching
Prior to this change, a capture group was being created for every
single character in the comment.

For any comments over 1024 characters (not uncommon when commenting out
a chunk of markup), PHP 7 would fail to match the regex due to hitting
a JIT stack limit.

Adding the `s` modifier tells the regex engine to include newline
characters when matching the dot character and allows us to remove the
inner capture group. This drastically reduces the amount of memory the
regex uses and keeps it well within the JIT stack limit for typical
uses.

This increases the limit for comment length up to approximately 1
million characters, instead 1024.