Add option for scoped Cache locks
Before this change out of order releases of the same cache lock could lead
to situation where client A acquired the lock, took longer than the timeout,
which made client B successfully acquire the lock. If A now finishes while B
is still holding the lock A will release the lock that does not belong to it.

This fix introduces a unique value that is written as the cache value to stop
A from deleting B's lock.