Fix mix exception handling
I noticed that these tests were uncommented so I did some investigating into why. It turns out that two prs referenced below were merged on either the 5.7 and 5.8 branches. I believe that due to a merge gone wrong some of the functionality from the PR sent to 5.7 was removed again causing the tests to fail (which is why they were commented out). I've re-added the changes and re-enable the commented out tests which now pass again.

5.7 PR: https://github.com/laravel/framework/pull/26431
5.8 PR: https://github.com/laravel/framework/pull/26289