Simplify the arguments passed to the before callbacks
This commit removes some legacy code that was laying around since #12305 was merged.

Before that PR, the arguments were spread out before being passed to the callbacks,
so the user and the ability were merged with the other arguments, before being passed to the callback.

Since then, the arguments are being passed as an array,
so instead of merging the user and the ability with the array of arguments, which is superfluous,
we can directly pass the user, the ability, and the arguments directly to the callback.

This commit doesn't change the existing functionality, and it keeps backwards compatibility.

I have also added some assertions to the tests, since the arguments in the before callbacks were never tested.