Make inequality validation fail on different types rather than 500
There are four types of inequality validators: lt, lte, gt, gte.
If these validators reference another attribute, the attribute
must be of the same type. Before this change, if the payload has
different types, a 500 error is given. However, the type is not
determine by the rule configuration, but rather the HTTP request.
Instead of throwing an error, just fail the validation.