Fix worker timeout handler for null `$job`
The queue worker timeout handler may be called with a null `$job` if the
timeout is reached when there is no job processing (perhaps it took too long
to fetch the next job in the worker loop). This fix checks to make sure there
is a job before attempting to mark the as failed if it will exceed the maximum
number of attempts.