search for similar results
first we take the first key/value pair of the expected data, and look to see if that exists in the database. If that finds results, we'll send those back to the developer since they're easier to compare.

if we don't find any of those results, we'll fallback to the old way of just returning results from the table.